pyproject.toml
[project]
name = "ibsys2scm"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.1",
    "cvxpy>=1.6.4",
    "fastapi>=0.115.12",
    "frozendict>=2.4.6",
    "jupyter>=1.1.1",
    "markdown-include>=0.8.1",
    "matplotlib>=3.10.1",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.7",
    "networkx>=3.4.2",
    "numpy>=2.2.3",
    "ortools>=9.12.4544",
    "pandas>=2.2.3",
    "psycopg[binary,pool]>=3.2.6",
    "pybase62>=1.0.0",
    "pydantic>=2.11.1",
    "pyomo[optional]>=6.9.1",
    "pytest>=8.3.5",
    "pyyaml>=6.0.2",
    "scipy>=1.15.2",
    "sqlalchemy>=2.0.39",
    "tabulate>=0.9.0",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.2",
]
[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
# Same as Black.
line-length = 100
indent-width = 4

# Assume Python 3.13
target-version = "py313"

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
ignore = ["F401"]
fixable = ["ALL"]
unfixable = ["F401"]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"
